+++ lib/dhttpd.dart
@@ -33,11 +33,13 @@ class Dhttpd {
     String? path,
     int port = defaultPort,
     Object address = defaultHost,
+    Map<String, String>? headers,
   }) async {
     path ??= Directory.current.path;
 
     final pipeline = const Pipeline()
         .addMiddleware(logRequests())
+        .addMiddleware(_headersMiddleware(headers))
         .addHandler(createStaticHandler(path, defaultDocument: 'index.html'));
 
     final server = await io.serve(pipeline, address, port);
@@ -46,3 +48,15 @@ class Dhttpd {
 
   Future<void> destroy() => _server.close();
 }
+
+Middleware _headersMiddleware(Map<String, String>? headers) =>
+    (Handler innerHandler) => (Request request) async {
+          final response = await innerHandler(request);
+          final responseHeaders = Map<String, String>.from(response.headers);
+          if (headers != null) {
+            for (var entry in headers.entries) {
+              responseHeaders[entry.key] = entry.value;
+            }
+          }
+          return response.change(headers: responseHeaders);
+        };